fix(bootlog): make the header public, and cast the uptime print
The sample app is a separate CMake target and does not have main/src on its
include path, so `#include "meshtastic_bootlog.h"` compiled everywhere in the
library and failed only when main.c reached it. Moved to
include/zephyr/meshtastic/bootlog.h alongside logring.h, which is where this
project already puts an API the sample consumes, and pulled in sys/util.h since
a public header cannot assume BIT() has arrived from somewhere else.

Also casts k_uptime_seconds() at the shell print: its width varies with the
timeout configuration, so a bare %lld is right on one build and a warning on
another.

+++ src/meshtastic_shell.c
@@ -8,6 +8,7 @@
 #include <string.h>
 
 #include <zephyr/kernel.h>
+#include <zephyr/meshtastic/bootlog.h>
 #include <zephyr/logging/log.h>
 #include <zephyr/shell/shell.h>
 #include <zephyr/version.h>
@@ -48,7 +49,6 @@
 #include "meshtastic_hlc.h"
 #include "meshtastic_config_store.h"
 #include "meshtastic_core.h"
-#include "meshtastic_bootlog.h"
 #include "meshtastic_preset.h"
 #include "meshtastic_region_presets.h"
 #include "meshtastic_powermon.h"
@@ -3194,7 +3194,7 @@ static int cmd_resets(const struct shell *sh, size_t argc, char **argv)
 		shell_print(sh, "cause       unavailable — register empty, or consumed by the "
 				"bootloader before the app ran");
 	}
-	shell_print(sh, "uptime      %lld s", k_uptime_seconds());
+	shell_print(sh, "uptime      %lld s", (long long)k_uptime_seconds());
 
 	n = meshtastic_bootlog_history(hist, ARRAY_SIZE(hist));
 	if (n == 0U) {
